Add files via upload
The Chatterbot_2 contains the python script which is used to program the chatbot. I have used PyQT5  to make the UI of the chatbot and also I have created the knowledge base file to train the bot on different questions. You are free to use and to modify this code for your personal projects.